Move Macros back and make it public
