+++ core/src/main/scala-3/caliban/schema/macros/Macros.scala
@@ -1,12 +1,10 @@
-package caliban.schema
+package caliban.schema.macros
 
 import caliban.schema.Annotations.GQLExcluded
-import caliban.schema.macros.TypeInfo
 
 import scala.quoted.*
-import scala.compiletime.*
 
-private object Macros {
+object Macros {
   // this code was inspired from WIP in magnolia
   // https://github.com/propensive/magnolia/blob/b937cf2c7dabebb8236e7e948f37a354777fa9b7/src/core/macro.scala
